TOTPTokenWidgetTile refactor: the countdown is now a single widget for better performance.
