+++ lib/views/main_view/main_view_widgets/token_widgets/totp_token_widgets/totp_token_widget_tile.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:flutterlifecyclehooks/flutterlifecyclehooks.dart';
 
 import '../../../../../l10n/app_localizations.dart';
 import '../../../../../model/tokens/totp_token.dart';
@@ -11,6 +10,7 @@ import '../../../../../widgets/custom_texts.dart';
 import '../../../../../widgets/custom_trailing.dart';
 import '../../../../../widgets/hideable_widget_.dart';
 import '../token_widget_tile.dart';
+import 'totp_token_widget_tile_countdown.dart';
 
 class TOTPTokenWidgetTile extends ConsumerStatefulWidget {
   final TOTPToken token;
@@ -22,11 +22,7 @@ class TOTPTokenWidgetTile extends ConsumerStatefulWidget {
   ConsumerState<TOTPTokenWidgetTile> createState() => _TOTPTokenWidgetTileState();
 }
 
-class _TOTPTokenWidgetTileState extends ConsumerState<TOTPTokenWidgetTile> with SingleTickerProviderStateMixin, LifecycleMixin {
-  double secondsLeft = 0;
-  late AnimationController animation;
-  late DateTime lastCount;
-
+class _TOTPTokenWidgetTileState extends ConsumerState<TOTPTokenWidgetTile> {
   void _copyOtpValue() {
     if (globalRef?.read(disableCopyOtpProvider) ?? false) return;
 
@@ -39,56 +35,6 @@ class _TOTPTokenWidgetTileState extends ConsumerState<TOTPTokenWidgetTile> with 
     );
     Future.delayed(const Duration(seconds: 5), () {
       globalRef?.read(disableCopyOtpProvider.notifier).state = false;
-    });
-  }
-
-  @override
-  void initState() {
-    super.initState();
-    animation = AnimationController(
-      vsync: this,
-      duration: Duration(seconds: widget.token.period),
-    );
-    animation.forward(from: 1 - widget.token.secondsUntilNextOTP / widget.token.period);
-    secondsLeft = widget.token.secondsUntilNextOTP;
-    lastCount = DateTime.now();
-    _startCountDown();
-  }
-
-  @override
-  dispose() {
-    animation.dispose();
-    super.dispose();
-  }
-
-  void _startCountDown() {
-    final now = DateTime.now();
-    final msSinceLastCount = now.difference(lastCount).inMilliseconds;
-    lastCount = now;
-    if (!mounted) return;
-    if (secondsLeft - (msSinceLastCount / 1000) > 0) {
-      setState(() => secondsLeft -= msSinceLastCount / 1000);
-    } else {
-      setState(() => secondsLeft = widget.token.secondsUntilNextOTP);
-      animation.forward(from: 1 - secondsLeft / widget.token.period);
-    }
-
-    final msUntilNextSecond = (secondsLeft * 1000).toInt() % 1000 + 1; // +1 to avoid 0
-    Future.delayed(Duration(milliseconds: msUntilNextSecond), () => _startCountDown());
-  }
-
-  @override
-  void onAppPause() {
-    if (!mounted) return;
-    animation.stop();
-  }
-
-  @override
-  void onAppResume() {
-    if (!mounted) return;
-    setState(() {
-      secondsLeft = widget.token.secondsUntilNextOTP;
-      animation.forward(from: 1 - secondsLeft / widget.token.period);
     });
   }
 
@@ -136,24 +82,7 @@ class _TOTPTokenWidgetTileState extends ConsumerState<TOTPTokenWidgetTile> with 
         child: HideableWidget(
           token: widget.token,
           isHidden: widget.token.isHidden && !widget.isPreview,
-          child: Stack(
-            alignment: Alignment.center,
-            children: [
-              Text(
-                '${secondsLeft.round()}',
-                overflow: TextOverflow.fade,
-                softWrap: false,
-              ),
-              AnimatedBuilder(
-                animation: animation,
-                builder: (context, child) {
-                  return CircularProgressIndicator(
-                    value: animation.value,
-                  );
-                },
-              ),
-            ],
-          ),
+          child: TotpTokenWidgetTileCountdown(widget.token),
         ),
       ),
     );
